inventory Add/Edit

+++ MPC.Web/Models/StockSubCategoryDropDown.cs
@@ -8,7 +8,7 @@
         /// <summary>
         /// Sub Category Id
         /// </summary>
-        public int SubCategoryId { get; set; }
+        public long SubCategoryId { get; set; }
 
         /// <summary>
         /// Name
@@ -18,6 +18,6 @@
         /// <summary>
         /// Category Id
         /// </summary>
-        public int CategoryId { get; set; }
+        public long CategoryId { get; set; }
     }
 }
